begin moving variables to global scope

+++ field/field.go
@@ -7,9 +7,9 @@ import (
 )
 
 type Field struct {
-    width int
+    width float64
 	cols int
-	colWidth int
+	colWidth float64
 	lineWidth float64
 	Imd *imdraw.IMDraw
 	color pixel.RGBA
@@ -19,20 +19,20 @@ func (f *Field) Draw() {
 	f.Imd.Reset()
 	f.Imd.Clear()
 	f.Imd.Color = f.color
-	for i := 0; i <= f.cols; i++ {
-		f.Imd.Push(pixel.V(float64(i * f.colWidth), 0), pixel.V(float64(i * f.colWidth), float64(f.width)))
+	for i := 0.0; i <= float64(f.cols); i++ {
+		f.Imd.Push(pixel.V(i * f.colWidth, 0), pixel.V(i * f.colWidth, f.width))
 		f.Imd.Line(f.lineWidth)
-		f.Imd.Push(pixel.V(0, float64(i * f.colWidth)), pixel.V(float64(f.width), float64(i * f.colWidth)))
+		f.Imd.Push(pixel.V(0, i * f.colWidth), pixel.V(f.width, i * f.colWidth))
 		f.Imd.Line(f.lineWidth)
 	}
 }
 
-func Create(width int, cols int, colWidth int, color pixel.RGBA) Field {
+func Create(width float64, cols int, colWidth float64, color pixel.RGBA) Field {
 	f := Field{}
 	f.width = width
 	f.cols = cols
 	f.colWidth = colWidth
-	f.lineWidth = math.Max(float64(colWidth) / 10, 2)
+	f.lineWidth = math.Max(colWidth / 10, 2)
 	f.Imd = imdraw.New(nil)
 	f.color = color
 	f.Draw()
